+++ CHANGELOG.md
@@ -34,3 +34,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `hb_font_funcs_t`, since we are using internal OpenType parser.
 - Most of the `hb_font_t` query API.
 - `hb_buffer_deserialize_glyphs`
+- `name` table support. Unused.